Create R markdown for project consolidated
My first pass at organizing everything into a markdown, feel free to add your code chunks for knn and model building so we can synchronize on everything. I also added a variable called "nurse to patient ratio" it's the number of nurse/number of daily patients. I think it takes care of the multicollinearity of the three variables but I'd be interested your thoughts.